# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 即刻AI-PPT API
  version: 1.0.0
  description: P1 contract baseline. Implementations are delivered by later Goals.
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /v1/drafts/{draftId}/generation-jobs:
    post:
      operationId: createGenerationJob
      tags:
        - drafts
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGenerationJobRequest"
      responses:
        "202":
          description: Successful response
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponse"
        default:
          description: RFC 7807 problem
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /v1/jobs/{jobId}:
    get:
      operationId: getGenerationJob
      tags:
        - jobs
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponse"
        default:
          description: RFC 7807 problem
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /v1/jobs/{jobId}/events:
    get:
      operationId: streamJobEvents
      tags:
        - jobs
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
      responses:
        "200":
          description: Successful response
          content:
            text/event-stream:
              schema:
                type: string
        default:
          description: RFC 7807 problem
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /v1/jobs/{jobId}:cancel:
    post:
      operationId: cancelGenerationJob
      tags:
        - jobs
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MutationRequest"
      responses:
        "202":
          description: Successful response
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponse"
        default:
          description: RFC 7807 problem
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /v1/jobs/{jobId}/slides/{slideId}:retry:
    post:
      operationId: retryGenerationSlide
      tags:
        - jobs
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
        - name: slideId
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MutationRequest"
      responses:
        "202":
          description: Successful response
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponse"
        default:
          description: RFC 7807 problem
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProblemDetails:
      title: ProblemDetails
      type: object
      additionalProperties: false
      properties:
        schemaVersion:
          type: integer
          const: 1
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        code:
          enum:
            - validation_error
            - authorization_error
            - quota_exceeded
            - unsafe_file
            - parse_failed
            - provider_rate_limited
            - provider_timeout
            - engine_contract_failed
            - slide_render_failed
            - slide_qa_failed
            - compile_failed
            - package_qa_failed
            - cancelled_by_user
            - idempotency_key_reused
            - revision_conflict
            - internal_error
        retryable:
          type: boolean
        requestId:
          type: string
          minLength: 1
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - field
              - message
            properties:
              field:
                type: string
                minLength: 1
              message:
                type: string
                minLength: 1
      required:
        - schemaVersion
        - type
        - title
        - status
        - detail
        - instance
        - code
        - retryable
        - requestId
        - fieldErrors
    MutationRequest:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - data
      properties:
        schemaVersion:
          type: integer
          const: 1
        data:
          type: object
          additionalProperties: true
        baseRevisionId:
          anyOf:
            - type: string
              pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
            - type: "null"
    GenerationImagePolicy:
      title: GenerationImagePolicy
      oneOf:
        - type: object
          additionalProperties: false
          required:
            - scope
            - usage
            - notes
          properties:
            scope:
              const: none
            usage:
              const:
                - none
            notes:
              type: object
              maxProperties: 0
            aiPath:
              type: "null"
            aiPathChain:
              type: array
              maxItems: 0
        - type: object
          additionalProperties: false
          required:
            - scope
            - usage
            - notes
            - aiPath
            - aiPathChain
          properties:
            scope:
              const: cover_only
            usage:
              const:
                - ai
            notes:
              type: object
              additionalProperties: false
              required:
                - cover
              properties:
                cover: &a1
                  type: string
                  minLength: 1
            aiPath:
              enum:
                - auto
                - api
                - host-native
                - manual
            aiPathChain:
              type: array
              minItems: 1
              maxItems: 3
              uniqueItems: true
              items:
                enum:
                  - api
                  - host-native
                  - manual
        - type: object
          additionalProperties: false
          required:
            - scope
            - usage
            - notes
            - aiPath
            - aiPathChain
          properties:
            scope:
              const: selective
            usage:
              const:
                - ai
            notes:
              type: object
              minProperties: 1
              propertyNames:
                pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
              additionalProperties: *a1
            aiPath:
              enum:
                - auto
                - api
                - host-native
                - manual
            aiPathChain:
              type: array
              minItems: 1
              maxItems: 3
              uniqueItems: true
              items:
                enum:
                  - api
                  - host-native
                  - manual
    CreateGenerationJobData:
      title: CreateGenerationJobData
      type: object
      additionalProperties: false
      properties:
        intentRevisionId:
          anyOf:
            - &a2
              type: string
              pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
            - type: "null"
        outlineRevisionId:
          anyOf:
            - *a2
            - type: "null"
        templateVersionId:
          anyOf:
            - *a2
            - type: "null"
        slideCount:
          type: integer
          minimum: 1
          maximum: 30
          default: 3
        sourceHashes:
          type: array
          items:
            type: string
            pattern: ^[a-f0-9]{64}$
        failureModes:
          type: object
          additionalProperties:
            enum:
              - none
              - once
              - always
        stepDelayMs:
          type: integer
          minimum: 0
          maximum: 10000
          default: 0
        crashOnceAtPosition:
          anyOf:
            - type: integer
              minimum: 1
              maximum: 30
            - type: "null"
        continueLimitedDraft:
          type: boolean
          default: false
        authorizeStrategistDesignLock:
          type: boolean
          default: false
          description: Explicitly authorizes the Strategist design proposal and spec lock before Executor authoring.
        imagePolicy:
          $ref: "#/components/schemas/GenerationImagePolicy"
    CreateGenerationJobRequest:
      title: CreateGenerationJobRequest
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - data
      properties:
        schemaVersion:
          type: integer
          const: 1
        data:
          $ref: "#/components/schemas/CreateGenerationJobData"
        baseRevisionId:
          anyOf:
            - *a2
            - type: "null"
    ResourceResponse:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - resourceId
        - resourceType
        - data
      properties:
        schemaVersion:
          type: integer
          const: 1
        resourceId:
          type: string
          pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
        resourceType:
          type: string
          minLength: 1
        data:
          type: object
          additionalProperties: true
        nextCursor:
          anyOf:
            - type: string
            - type: "null"